# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurante API",
        "description": "Documentación del backend (Laravel + Swagger).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/menu-items": {
            "get": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Listar ítems del menú (paginado)",
                "operationId": "f8f6266a0f281d4a6fd2924521990d27",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Limonada"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Natural",
                                                        "nullable": true
                                                    },
                                                    "categoria": {
                                                        "type": "string",
                                                        "example": "bebida"
                                                    },
                                                    "precio": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 6.5
                                                    },
                                                    "imagen": {
                                                        "type": "string",
                                                        "example": "limonada.jpg",
                                                        "nullable": true
                                                    },
                                                    "disponible": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Crear ítem de menú",
                "operationId": "080371a7ab243e7fe3fb2e21af1190a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "categoria",
                                    "precio"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Limonada"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Natural",
                                        "nullable": true
                                    },
                                    "categoria": {
                                        "type": "string",
                                        "example": "bebida"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 6.5
                                    },
                                    "imagen": {
                                        "type": "string",
                                        "example": "limonada.jpg",
                                        "nullable": true
                                    },
                                    "disponible": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            }
        },
        "/api/menu-items/{id}": {
            "get": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Ver ítem",
                "operationId": "9a846f4be41fcdbe9832b94b39f95bf4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Actualizar ítem",
                "operationId": "65192842ac27acc7bbe511a3e3d42215",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Limonada natural"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Con hierbabuena",
                                        "nullable": true
                                    },
                                    "categoria": {
                                        "type": "string",
                                        "example": "bebida"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 7
                                    },
                                    "imagen": {
                                        "type": "string",
                                        "example": "limonada_2.jpg",
                                        "nullable": true
                                    },
                                    "disponible": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Eliminar ítem",
                "operationId": "d28acee26f2a00a051b272f4b60af611",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
    },
...
}